chore: Moved IRoute to src/types/models
