+++ src/api/route.service.ts
@@ -1,5 +1,5 @@
 import api from "@api/index";
-import { IRoute } from "src/types/route";
+import { IRoute } from "src/types/models/route";
 
 export const getRoutes = async (): Promise<IRoute[]> => {
   return api.get("/routes");
